Simple bot using Microsoft bot framework,LUIS and google custom search api
A Simple bot using Microsoft bot framework,LUIS and google custom search
api.
A databse of countries and capitals is also mantained , currently there
are only two countries and their capitals stored in that.
User can tell bot to search on any topic using any statement(english)
and bot will provide you with the title and the link of the webpage.